# Conflicts:
#	src/instructions/v.rs

+++ definitions/src/lib.rs
@@ -22,5 +22,5 @@ pub const ISA_B: u8 = 0b0000_0001;
 pub const ISA_MOP: u8 = 0b0000_0010;
 pub const ISA_V: u8 = 0b0000_0100;
 
-pub const VLEN: u32 = 2048;
-pub const ELEN: u32 = 1024;
+pub const VLEN: usize = 2048;
+pub const ELEN: usize = 1024;
